Create API Quiz endpoints #3
This commit covers all of Chapter 1's requested features, at a price.
I won't consider it closed until all required unit and e2e tests are
provided (next step).

There are some specific choices (commented within the code) that I'm
not happy with, but I refrained from coding this the "proper way" to
avoid overshooting. Starting with a non-persistent app, with no ORM,
and no proper entities... most of the validation will come out of the
box once their relationships are set via many-to-one or many-to-many
clauses in Chapter 2, so it makes sense to leave the "proper way" for
that refactor.

And I would have gone some extra miles if Chapter 2's refactor wasn't
just around the corner. But, that's the kind of thing I would debate
with the team before actual coding work.

Work Items Resolved:
https://github.com/amadurell/nest-quiz-me/issues/3

Note: I've actually split that issue in 2 halves, moving the 2nd one
to Chapter 2:
https://github.com/amadurell/nest-quiz-me/issues/12