Tests #5
Tests for the quizzes service, in light of the upcoming refactor, are
complete. Tests for the controller should also be implemented, mainly
to cover cases that are caught by the validation pipe, which doesn't
let a wrong input get to the service.

For sake of timing I will add those after the database refactor, when
many pieces will be moved to each corresponding entity.

Work Items Resolved:
https://github.com/amadurell/nest-quiz-me/issues/5