Just a quick README.md update
Besides having added/modified some scripts within the package.json,
I've realized the instructions referred to npm as the package manager
when I've been using pnpm. It would still run, but it might complain
about the lock file being an old version or something (which is not
the case, it's just... pnpm).