Features/chapter1 (#13)
* Create entities and Dtos

I've run the following instructions on the command line:
    - nest g res Quizzes
    - nest g res Quizzes/Questions --flat
    - nest g res Quizzes/Answers --flat
    - pnpm add class-validator class-transformer

Then have added the properties to each entity and each dto, readonly
modifiers to dto properties, with corresponding basic validation as
per each type. QuizzesModule has been added automatically to the app
module class, and I have added the ValidationPipe manually in the
bootstrap function in main.ts

Work Items Resolved:
https://github.com/amadurell/nest-quiz-me/issues/2

* Create entities and Dtos

I've run the following instructions on the command line:
    - nest g res Quizzes
    - nest g res Quizzes/Questions --flat
    - nest g res Quizzes/Answers --flat
    - pnpm add class-validator class-transformer

Then have added the properties to each entity and each dto, readonly
modifiers to dto properties, with corresponding basic validation as
per each type. QuizzesModule has been added automatically to the app
module class, and I have added the ValidationPipe manually in the
bootstrap function in main.ts

The rest of the files are just a blueprint, but I'm adding them here
so it won't break compilation.

Work Items Resolved:
https://github.com/amadurell/nest-quiz-me/issues/2

Work Items in progress (which won't break compilation)
https://github.com/amadurell/nest-quiz-me/issues/3

* Create API Quiz endpoints #3

This commit covers all of Chapter 1's requested features, at a price.
I won't consider it closed until all required unit and e2e tests are
provided (next step).

There are some specific choices (commented within the code) that I'm
not happy with, but I refrained from coding this the "proper way" to
avoid overshooting. Starting with a non-persistent app, with no ORM,
and no proper entities... most of the validation will come out of the
box once their relationships are set via many-to-one or many-to-many
clauses in Chapter 2, so it makes sense to leave the "proper way" for
that refactor.

And I would have gone some extra miles if Chapter 2's refactor wasn't
just around the corner. But, that's the kind of thing I would debate
with the team before actual coding work.

Work Items Resolved:
https://github.com/amadurell/nest-quiz-me/issues/3

Note: I've actually split that issue in 2 halves, moving the 2nd one
to Chapter 2:
https://github.com/amadurell/nest-quiz-me/issues/12

* Tests #5

Tests for the quizzes service, in light of the upcoming refactor, are
complete. Tests for the controller should also be implemented, mainly
to cover cases that are caught by the validation pipe, which doesn't
let a wrong input get to the service.

For sake of timing I will add those after the database refactor, when
many pieces will be moved to each corresponding entity.

Work Items Resolved:
https://github.com/amadurell/nest-quiz-me/issues/5

* Just a quick README.md update

Besides having added/modified some scripts within the package.json,
I've realized the instructions referred to npm as the package manager
when I've been using pnpm. It would still run, but it might complain
about the lock file being an old version or something (which is not
the case, it's just... pnpm).

Co-authored-by: amadurell <alexandre.madurelldeotero.adicomgroup@skytv.it>