Refactor app to use the postgredb DB #7
The application now makes use of TypeORM and Entity features, so most
(well... all) methods in the Repository classes are commented out, so
as to keep the previous Exception messages for comparison, in view of
the required work in the unit tests files.

It's taken me quite a while longer than expected, some of the config
parameters and decorators for validation kept causing strange errors.
Had to try different combinations of whitelist, forbidNonWhitelisted
and transform in the validation pipe (main.ts) and different methods
to get the parent entities to cascade to their children on creation,
but it still wouldn't work properly until I removed @ValidateNested
from the questions and answers properties of the dtos.

Work Items Resolved:
https://github.com/amadurell/nest-quiz-me/issues/7